+++ inst/examples/PPT.FitGraphicIntoShapeExample.R
@@ -5,8 +5,8 @@ p <- PPT.Init(visible=T, method = "RDCOMClient")
 p <- PPT.AddBlankSlide(p)
 
 # add two rectangle shapes to slide
-p <- PPT.AddRectangle(p, width = .4, left=.1)                      # add a shape to slide
-p <- PPT.AddRectangle(p, width = .4, left=.5, height=.45, top=.5)  # add a shape to slide
+p <- PPT.AddShape(p, width = .4, left=.1)                      # add a shape to slide
+p <- PPT.AddShape(p, width = .4, left=.5, height=.45, top=.5)  # add a shape to slide
 
 # get all shapes on current slide and replace by image
 # shapes are not destroyed to see the image placement